Added support for dimensions to metrics

+++ modules/CheckDisk/check_drive.cpp
@@ -288,7 +288,7 @@ parsers::where::node_type calculate_total_used(boost::shared_ptr<filter_obj> obj
 	} else {
 		number = str::format::decode_byte_units(number, unit);
 	}
-	return parsers::where::factory::create_int(number);
+	return parsers::where::factory::create_int(static_cast<long long>(number));
 }
 
 parsers::where::node_type calculate_user_used(boost::shared_ptr<filter_obj> object, parsers::where::evaluation_context context, parsers::where::node_type subject) {
@@ -301,7 +301,7 @@ parsers::where::node_type calculate_user_used(boost::shared_ptr<filter_obj> obje
 	} else {
 		number = str::format::decode_byte_units(number, unit);
 	}
-	return parsers::where::factory::create_int(number);
+	return parsers::where::factory::create_int(static_cast<long long>(number));
 }
 int do_convert_type(const std::string &keyword) {
 	if (keyword == "fixed")
@@ -504,7 +504,7 @@ public:
 		hlp::tchar_buffer fileSysName(1024);
 		DWORD maximumComponentLength, fileSystemFlags;
 
-		if (!ptrGetVolumeInformationByHandleW(hVolume, volumeName.get(), volumeName.size(),
+		if (!ptrGetVolumeInformationByHandleW(hVolume, volumeName.get(), static_cast<DWORD>(volumeName.size()),
 			NULL, &maximumComponentLength, &fileSystemFlags, fileSysName.get(), static_cast<DWORD>(fileSysName.size()))) {
 			NSC_LOG_ERROR("Failed to get volume information: " + error::lookup::last_error());
 		} else {
